refactor: move workflow providers to workflow_builder\providers
