+++ packages/api_testing_suite/lib/src/workflow_builder/providers/workflow_providers.dart
@@ -1,175 +1,188 @@
+import 'package:api_testing_suite/api_testing_suite.dart';
+import 'package:api_testing_suite/src/workflow_builder/models/workflow_log_entry.dart' as workflow_log;
+import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../models/models.dart';
 import 'workflows_notifier.dart';
 
-final workflowsProvider = StateNotifierProvider<WorkflowsNotifier, List<WorkflowModel>>((ref) {
-  return WorkflowsNotifier(ref);
+final workflowsNotifierProvider =
+    StateNotifierProvider<WorkflowsNotifier, List<WorkflowModel>>((ref) {
+  return WorkflowsNotifier();
 });
 
 final currentWorkflowProvider = StateProvider<String?>((ref) => null);
 
+final selectedNodeIdProvider = StateProvider<String?>((ref) => null);
+
 final activeWorkflowProvider = Provider<WorkflowModel?>((ref) {
   final currentWorkflowId = ref.watch(currentWorkflowProvider);
-  final workflows = ref.watch(workflowsProvider);
-  
+  final workflows = ref.watch(workflowsNotifierProvider);
+
   if (currentWorkflowId == null) return null;
-  
+
   return workflows.firstWhere(
     (workflow) => workflow.id == currentWorkflowId,
-    orElse: () => WorkflowModel.create(name: 'Workflow $currentWorkflowId'),
+    orElse: () => throw Exception('Workflow not found'),
   );
 });
 
-final workflowExecutionStateProvider = StateNotifierProvider<WorkflowExecutionNotifier, WorkflowExecutionState>((ref) {
-  return WorkflowExecutionNotifier(ref);
+final workflowExecutionStateProvider =
+    StateProvider<WorkflowExecutionState>((ref) {
+  return const WorkflowExecutionState();
 });
 
-enum WorkflowExecutionStatus {
-  idle,
-  running,
-  paused,
-  completed,
-}
+final connectionModeProvider = StateProvider<bool>((ref) => false);
 
-class WorkflowExecutionState {
-  final WorkflowExecutionStatus status;
-  final String? currentNodeId;
-  final List<String> executedNodeIds;
-  final Map<String, dynamic> executionContext;
-
-  WorkflowExecutionState({
-    this.status = WorkflowExecutionStatus.idle,
-    this.currentNodeId,
-    this.executedNodeIds = const [],
-    this.executionContext = const {},
-  });
-
-  WorkflowExecutionState copyWith({
-    WorkflowExecutionStatus? status,
-    String? currentNodeId,
-    List<String>? executedNodeIds,
-    Map<String, dynamic>? executionContext,
-  }) {
-    return WorkflowExecutionState(
-      status: status ?? this.status,
-      currentNodeId: currentNodeId ?? this.currentNodeId,
-      executedNodeIds: executedNodeIds ?? this.executedNodeIds,
-      executionContext: executionContext ?? this.executionContext,
-    );
-  }
-}
-
-class WorkflowExecutionNotifier extends StateNotifier<WorkflowExecutionState> {
-  final Ref _ref;
+final workflowLogsProvider = Provider.family<List<workflow_log.WorkflowLogEntry>, String>((ref, workflowId) {
+  final workflows = ref.watch(workflowsNotifierProvider);
+  final workflowModel = workflows.firstWhere(
+    (workflow) => workflow.id == workflowId,
+    orElse: () => throw Exception('Workflow not found'),
+  );
   
-  WorkflowExecutionNotifier(this._ref) : super(WorkflowExecutionState());
-
-  Future<void> start() async {
-    if (state.status == WorkflowExecutionStatus.running) return;
-    
-    final workflow = _ref.read(activeWorkflowProvider);
-    if (workflow == null || workflow.nodes.isEmpty) return;
-
-    final firstNode = workflow.nodes.first;
-    
-    state = WorkflowExecutionState(
-      status: WorkflowExecutionStatus.running,
-      currentNodeId: firstNode.id,
-      executedNodeIds: [],
-      executionContext: {},
+  final executionState = ref.watch(workflowExecutionStateProvider);
+  
+  final logs = <workflow_log.WorkflowLogEntry>[];
+  
+  logs.add(
+    workflow_log.WorkflowLogEntry(
+      message: 'Workflow initialized',
+      level: LogLevel.info,
+      timestamp: workflowModel.createdAt,
+    ),
+  );
+  
+  if (executionState.startTime != null) {
+    logs.add(
+      workflow_log.WorkflowLogEntry(
+        message: 'Execution started',
+        level: LogLevel.info,
+        timestamp: executionState.startTime,
+      ),
     );
-
-    await _executeNode(firstNode.id);
-  }
-
-  Future<void> pause() async {
-    if (state.status != WorkflowExecutionStatus.running) return;
     
-    state = state.copyWith(status: WorkflowExecutionStatus.paused);
-  }
-
-  Future<void> resume() async {
-    if (state.status != WorkflowExecutionStatus.paused) return;
-    
-    state = state.copyWith(status: WorkflowExecutionStatus.running);
-    
-    if (state.currentNodeId != null) {
-      await _executeNode(state.currentNodeId!);
+    for (final nodeId in executionState.executedNodeIds) {
+      final node = workflowModel.nodes.firstWhere(
+        (n) => n.id == nodeId,
+        orElse: () => throw Exception('Node not found'),
+      );
+      
+      final result = executionState.executionResults[nodeId];
+      final success = result != null && result['status'] == 'success';
+      
+      logs.add(
+        workflow_log.WorkflowLogEntry(
+          message: success 
+              ? 'Node execution completed successfully' 
+              : 'Node execution failed',
+          level: success ? LogLevel.success : LogLevel.error,
+          nodeId: nodeId,
+          details: node.nodeType == NodeType.request
+              ? 'URL: ${node.requestModel?.url ?? "No URL"}\nMethod: ${node.requestModel?.method ?? "GET"}'
+              : null,
+        ),
+      );
     }
-  }
-
-  Future<void> stop() async {
-    state = WorkflowExecutionState();
     
-    final workflow = _ref.read(activeWorkflowProvider);
-    if (workflow != null) {
-      _ref.read(workflowsProvider.notifier).updateNodes(
-        workflow.id,
-        workflow.nodes.map((node) => node.copyWith(status: NodeStatus.inactive)).toList(),
+    if (executionState.currentNodeId != null) {
+      logs.add(
+        workflow_log.WorkflowLogEntry(
+          message: 'Node execution in progress',
+          level: LogLevel.info,
+          nodeId: executionState.currentNodeId,
+        ),
+      );
+    }
+    
+    if (executionState.isCompleted && executionState.endTime != null) {
+      logs.add(
+        workflow_log.WorkflowLogEntry(
+          message: 'Workflow execution completed',
+          level: LogLevel.success,
+          timestamp: executionState.endTime,
+        ),
+      );
+    }
+    
+    if (executionState.isPaused) {
+      logs.add(
+        workflow_log.WorkflowLogEntry(
+          message: 'Workflow execution paused',
+          level: LogLevel.warning,
+        ),
+      );
+    }
+    
+    if (executionState.hasError) {
+      logs.add(
+        workflow_log.WorkflowLogEntry(
+          message: 'Workflow execution failed: ${executionState.errorMessage}',
+          level: LogLevel.error,
+        ),
       );
     }
   }
+  
+  return logs;
+});
 
-  Future<void> _executeNode(String nodeId) async {
-    if (state.status != WorkflowExecutionStatus.running) return;
-    
-    final workflow = _ref.read(activeWorkflowProvider);
-    if (workflow == null) return;
-    
-    final nodeIndex = workflow.nodes.indexWhere((node) => node.id == nodeId);
-    if (nodeIndex == -1) return;
-    
-    final node = workflow.nodes[nodeIndex];
-    
-    _ref.read(workflowsProvider.notifier).updateNode(
-      workflow.id,
-      node.copyWith(status: NodeStatus.running),
-    );
-    
-    await Future.delayed(const Duration(seconds: 1));
-    
-    final isSuccess = DateTime.now().millisecondsSinceEpoch % 5 != 0; 
-    
-    _ref.read(workflowsProvider.notifier).updateNode(
-      workflow.id,
-      node.copyWith(status: isSuccess ? NodeStatus.success : NodeStatus.failure),
-    );
-    
-    final updatedExecutedNodeIds = [...state.executedNodeIds, nodeId];
-    state = state.copyWith(
-      executedNodeIds: updatedExecutedNodeIds,
-    );
-    
-    if (!isSuccess && node.connectedToIds.isNotEmpty) {
-      final connections = workflow.connections.where(
-        (conn) => conn.sourceId == nodeId && conn.isConditional
-      ).toList();
-      
-      if (connections.isNotEmpty) {
-        state = state.copyWith(status: WorkflowExecutionStatus.completed);
-        return;
+final workflowExecutionControlProvider =
+    Provider<WorkflowExecutionControl>((ref) {
+  final notifier = ref.watch(workflowsNotifierProvider.notifier);
+  final currentId = ref.watch(currentWorkflowProvider);
+
+  return WorkflowExecutionControl(
+    startExecution: () {
+      if (currentId != null) {
+        notifier.startExecution(
+            currentId,
+            (state) => ref.read(workflowExecutionStateProvider.notifier).state =
+                state);
       }
-    }
-    
-    final nextNodeIds = node.connectedToIds.isEmpty 
-        ? [] 
-        : workflow.connections
-            .where((conn) => conn.sourceId == nodeId)
-            .map((conn) => conn.targetId)
-            .toList();
-    
-    if (nextNodeIds.isEmpty) {
-      state = state.copyWith(status: WorkflowExecutionStatus.completed);
-      return;
-    }
-    
-    for (final nextNodeId in nextNodeIds) {
-      state = state.copyWith(currentNodeId: nextNodeId);
-      await _executeNode(nextNodeId);
-      
-      if (state.status != WorkflowExecutionStatus.running) {
-        break;
+    },
+    pauseExecution: () {
+      if (currentId != null) {
+        notifier.pauseExecution(currentId);
       }
-    }
+    },
+    resumeExecution: () {
+      if (currentId != null) {
+        notifier.resumeExecution(currentId);
+      }
+    },
+    stopExecution: () {
+      if (currentId != null) {
+        notifier.stopExecution(currentId);
+      }
+    },
+  );
+});
+
+final workflowExecutionLogsProvider = StateNotifierProvider<WorkflowExecutionLogsNotifier, List<LogEntry>>((ref) {
+  return WorkflowExecutionLogsNotifier();
+});
+
+class WorkflowExecutionLogsNotifier extends StateNotifier<List<LogEntry>> {
+  WorkflowExecutionLogsNotifier() : super([]);
+
+  void clearLogs() {
+    state = [];
   }
+
+  void addLog(LogEntry entry) {
+    state = [...state, entry];
+  }
+}
+
+class WorkflowExecutionControl {
+  final VoidCallback startExecution;
+  final VoidCallback pauseExecution;
+  final VoidCallback resumeExecution;
+  final VoidCallback stopExecution;
+
+  const WorkflowExecutionControl({
+    required this.startExecution,
+    required this.pauseExecution,
+    required this.resumeExecution,
+    required this.stopExecution,
+  });
 }
